ban params too with a callback
