+++ .importlinter
@@ -8,6 +8,7 @@ type = forbidden
 source_modules =
     openpilot.selfdrive.car
 forbidden_modules =
+    openpilot.common.params
     openpilot.system
     openpilot.body
     openpilot.docs
@@ -34,4 +35,9 @@ ignore_imports =
     openpilot.selfdrive.car.tests.test_car_interfaces -> openpilot.selfdrive.controls.lib.longcontrol
     openpilot.selfdrive.car.tests.test_car_interfaces -> openpilot.selfdrive.controls.lib.latcontrol_torque
     openpilot.selfdrive.car.tests.test_car_interfaces -> openpilot.selfdrive.controls.lib.latcontrol_pid
+    openpilot.selfdrive.car.card -> openpilot.common.params
+    openpilot.selfdrive.car.tests.test_models -> openpilot.common.params
+    # these two will still live in openpilot, but require some modification
+    openpilot.selfdrive.car.fw_versions -> openpilot.common.params
+    openpilot.selfdrive.car.ecu_addrs -> openpilot.common.params
 unmatched_ignore_imports_alerting = warn
