+++ system/hardware/tici/tests/test_power_draw.py
@@ -8,7 +8,8 @@ from tabulate import tabulate
 import cereal.messaging as messaging
 from cereal.services import SERVICE_LIST
 from openpilot.common.mock import mock_messages
-from openpilot.selfdrive.car.car_helpers import write_car_param
+from openpilot.common.params import Params
+from openpilot.selfdrive.car.car_helpers import get_demo_car_params
 from openpilot.system.hardware.tici.power_monitor import get_power
 from openpilot.system.manager.process_config import managed_processes
 from openpilot.system.manager.manager import manager_cleanup
@@ -41,7 +42,7 @@ PROCS = [
 class TestPowerDraw:
 
   def setup_method(self):
-    write_car_param()
+    Params().put("CarParams", get_demo_car_params().to_bytes())
 
     # wait a bit for power save to disable
     time.sleep(5)
